[infra] Remove hal.hpp and kernel_types.hpp from rtoptions.hpp include chain
rtoptions.hpp was pulling in hal.hpp (avg 829ms/TU parse cost, 157 TU
footprint) transitively, making it expensive to include.

Changes:
- Move HalProcessorSet from hal.hpp to hal_types.hpp (lightweight, already
  included by rtoptions.hpp via the hal_types path)
- Remove #include "llrt/hal.hpp" from rtoptions.hpp; replace with
  #include <tt-metalium/hal_types.hpp> and a forward decl: class Hal;
- Move ParseAllFeatureEnv inline body to rtoptions.cpp (was forcing
  rtoptions.hpp includers to parse the full Hal interface)
- Remove #include <tt-metalium/kernel_types.hpp>; forward-declare
  enum class KernelBuildOptLevel : uint8_t instead
- Add #include <mutex> for std::once_flag (was provided transitively via hal.hpp)
- Add explicit #include "llrt/hal.hpp" to metal_context.hpp (it already owned
  the Hal singleton via MetalContext::hal(); make the implicit dep explicit)

CBA result: hal.hpp and rtoptions.hpp both dropped off the top-10 expensive
headers list (previously hal.hpp: 130,209ms / 157 TUs / avg 829ms).
